moves version script to setup, tests update
